median filter - very slow though

+++ DisplayParameters.py
@@ -36,7 +36,7 @@ class DisplayPamrameters:
 		hsvM[:,:,2] = pv / 255.0
 		
 		self.HSVmap = matplotlib.colors.hsv_to_rgb(hsvM)
-		self.HSVmap = (self.HSVmap * 255.0).astype(np.int)
+		self.HSVmap = (self.HSVmap * 255.0).astype(np.uint32)
 		self.HSVmap = (255 << 24 | self.HSVmap[:,:,0] << 16 | self.HSVmap[:,:,1] << 8 | self.HSVmap[:,:,2]).flatten() 
 		self.HSVmap = self.HSVmap.astype(np.uint32)
